moves assessments to resource list #277
