+++ cms/urls.py
@@ -11,10 +11,9 @@ from wagtail.wagtaildocs import urls as wagtaildocs_urls
 
 from likes.views import save_like
 from resources.views import get_json_data
-from home.views import (
-    landing_page_controller,
-    assessment_controller,
-    assessment_summary_controller
+from home.views import landing_page_controller
+from resources.views import (
+    assessment_controller, assessment_summary_controller
 )
 
 from django.views.generic import TemplateView
@@ -45,7 +44,6 @@ urlpatterns = [
     url('sleep-and-stress/', landing_page_controller),
     url('sleep-and-mind-racing/', landing_page_controller),
     url('sleep/talk-about-it/', landing_page_controller),
-    url('assessment/', landing_page_controller),
 
     # For anything not caught by a more specific rule above, hand over to
     # Wagtail's page serving mechanism. This should be the last pattern in
