--- home/templates/home/assessment.html
@@ -1,103 +1,0 @@
-<!DOCTYPE html>
-<!--[if lt IE 7]>      <html class="no-js lt-ie9 lt-ie8 lt-ie7"> <![endif]-->
-<!--[if IE 7]>         <html class="no-js lt-ie9 lt-ie8"> <![endif]-->
-<!--[if IE 8]>         <html class="no-js lt-ie9"> <![endif]-->
-<!--[if gt IE 8]><!--> <html class="no-js"> <!--<![endif]-->
-    <head>
-      <!-- Google Tag Manager -->
-        <script>
-          dataLayer = [];
-        </script>
-        <script>(function(w,d,s,l,i){w[l]=w[l]||[];w[l].push({'gtm.start':
-        new Date().getTime(),event:'gtm.js'});var f=d.getElementsByTagName(s)[0],
-        j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src=
-        'https://www.googletagmanager.com/gtm.js?id='+i+dl;f.parentNode.insertBefore(j,f);
-        })(window,document,'script','dataLayer','GTM-PF8GGT7');</script>
-        <!-- End Google Tag Manager -->
-        <script>
-          (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
-          (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
-          m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
-          })(window,document,'script','https://www.google-analytics.com/analytics.js','ga');
-
-          ga('create', 'UA-97311089-1', 'auto');
-
-        </script>
-        <meta charset="utf-8" />
-        <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-        <title>
-
-                Coming Soon
-
-
-
-
-
-
-        </title>
-        <meta name="description" content="" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-
-
-
-          <link rel="stylesheet" href="https://npmcdn.com/tachyons-word-break@3.0.0/css/tachyons-word-break.min.css" />
-          <link rel="stylesheet" href="https://unpkg.com/tachyons/css/tachyons.min.css">
-          <link href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" rel="stylesheet" integrity="sha384-wvfXpqpZZVQGK6TAh5PVlGOfQNHSoD2xbE+QkPxCAFlNEevoEH3Sl0sibVcOQVnN" crossorigin="anonymous">
-          <!-- Taken from https://jonsuh.com/hamburgers/ -->
-          <link href="https://cdnjs.cloudflare.com/ajax/libs/hamburgers/0.9.1/hamburgers.min.css" rel="stylesheet">
-          <link rel="icon" href="/static/favicon.png">
-
-        <link href="/static/cms.css" rel="stylesheet" type="text/css" />
-        <script src="//apps.expert-24.com/WebBuilder/TraversalService/Embed2.js?algoid=4648&id=traversal-container&memberid=new&staticurl=true" type="text/javascript" defer></script>
-    </head>
-
-    <body class="template-homepage">
-      <!-- Google Tag Manager (noscript) -->
-      <noscript><iframe src="https://www.googletagmanager.com/ns.html?id=GTM-PF8GGT7"
-      height="0" width="0" style="display:none;visibility:hidden"></iframe></noscript>
-      <!-- End Google Tag Manager (noscript) -->
-
-<div class="lm-bg-white pa3 w-100 f6 f5-l f5-m banner segoe fixed top-0 z-5 dt-ns dt--fixed-ns shadow-2" id="beta-banner">
-  <div class="dtc-ns w-50-l w-40-m v-mid">
-    <div class="rich-text"><p><b>BETA: </b>This is a prototype - your <a>feedback</a> will help to improve it. <a href="http://localhost:8000/#alpha">Find out more</a></p></div>
-  </div>
-  <div class="dtc-ns dn tr v-mid banner_buttons">
-    <a href="/coming-soon" class="f6 f7-m f5 link dib ph3 pv2 br1 pointer nunito tracked inner-shadow-active lm-bg-dark-blue-hover lm-dark-blue lm-white-hover ba b--lm-dark-blue" id="assessment">Take self-assessment</a>
-    <a href="/crisis" class="f6 f7-m f5 link dib ph3 pv2 br1 pointer nunito tracked inner-shadow-active lm-bg-dark-blue-hover lm-dark-blue lm-white-hover ba b--lm-dark-blue crisis-link">Get urgent support</a>
-  </div>
-</div>
-
-  <section class="pt5-plus ph4 ph5-m ph6-l tc lm-white cover lh-copy bg-center">
-
-
-<a href="/">
-  <img src="/static/images/london_minds_logo.png" class="w-60 w-15-ns" />
-</a>
-
-  </section>
-  <section class="pt5-plus ph4 ph5-m ph6-l" id="traversal-container">
-    <div class="tc">
-      <div class="block-paragraph"><div class="rich-text"><h2>This Page is Coming Soon</h2><p><img><br/></p><p>We are in the alpha stages of development so some things are still being built.<br/></p><p>Please share any <a href="/#alpha">feedback</a> with us so we can improve the site.<br/></p></div></div>
-    </div>
-  </section>
-  <footer class="lm-bg-dark-blue tc pa3 f6 f5-ns tracked">
-    <div class="ph2 ph5-plus-ns lm-white">
-      <div class="rich-text"><p><a href="/terms-and-conditions/">Terms and Conditions</a> <a href="/resources/headspace/">Contact Us</a></p></div>
-    </div>
-  </footer>
-
-
-
-        <script type="text/javascript" src="/static/scripts.js" charset="utf-8"></script>
-
-
-
-
-
-        <div id="cookie_accept_bar" class="lm-bg-white tc pv3 dn bottom-0 fixed z-5 w-100">
-          <p>Before browsing our site, please accept our <a href="/cookie-policy">cookies policy</a></p>
-          <button type="submit"  class="f5 link dib ph3 pv2 br1 pointer nunito tracked inner-shadow-active lm-white lm-bg-dark-blue button lm-bg-orange-hover lm-dark-blue-hover mv3">ACCEPT AND CLOSE</button>
-
-        </div>
-    </body>
-</html>
