+++ resources/templates/resources/assessment.html
@@ -1,0 +1,18 @@
+{% extends "base.html" %}
+
+{% block content %}
+  {% if banner %}
+    {% include "resources/banner.html" with banner=banner %}
+  {% endif %}
+
+  <div class="tc w-100 mt6">
+    {% include "resources/logo.html" %}
+  </div>
+
+  <div class="mt3 pa5">
+    <div id="traversal-container"></div>
+    <script src="//apps.expert-24.com/WebBuilder/TraversalService/Embed2.js?algoid=4648&id=traversal-container&memberid=new&staticurl=true" type="text/javascript" defer></script>
+  </div>
+
+  {% include "resources/components/footer.html" with footer_blocks=footer_blocks footer_links=footer_links %}
+{% endblock %}
