+++ resources/templates/resources/assessment/bullets.html
